functions/blog_post_image_updater/requirements.txt
PyGithub==1.59.1
toml==0.10.2
click==8.0.3